refactor: 토큰 관련 로직 api 모듈로 옮김
