+++ growthmate-api/build.gradle
@@ -10,6 +10,10 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
     implementation 'org.springframework.boot:spring-boot-starter-web'
 
+    implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
+    runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.11.5'
+    runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.11.5'
+
     implementation 'com.querydsl:querydsl-jpa:5.0.0:jakarta'
 
     implementation 'org.springdoc:springdoc-openapi-ui:1.6.15' // swagger
